Refactor | HermitBullet 로직 CurveBullet에 적용
전반적인 메서드명도 손봤습니다.
