+++ Assets/03. Mingyu/01. Scene/Test_Mingyu_BossScene.unity
@@ -1155,7 +1155,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _isDontDestroy: 0
-  _bossPrefab: {fileID: 1875926273}
+  _bossController: {fileID: 1875926274}
 --- !u!1 &1543345149
 GameObject:
   m_ObjectHideFlags: 0
@@ -1596,6 +1596,17 @@ GameObject:
   m_CorrespondingSourceObject: {fileID: 4865756857718749494, guid: 71172bf5418bf194090c5e44d8c00fe4, type: 3}
   m_PrefabInstance: {fileID: 224139032}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &1875926274 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 2089149400775240186, guid: 71172bf5418bf194090c5e44d8c00fe4, type: 3}
+  m_PrefabInstance: {fileID: 224139032}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1875926273}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 56d1fb783479238428c5a248b36fd51a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1946357161
 GameObject:
   m_ObjectHideFlags: 0
