+++ Assets/03. Mingyu/03. Prefabs/Boss/P_Boss.prefab
@@ -266,7 +266,7 @@ Transform:
   m_GameObject: {fileID: 8750473511153552226}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 1.2, z: 0}
+  m_LocalPosition: {x: 0, y: 1.6, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
