+++ Assets/03. Mingyu/03. Prefabs/Boss/P_BulletHermitCurve_Boss.prefab
@@ -12,9 +12,9 @@ GameObject:
   - component: {fileID: 874983024548535551}
   - component: {fileID: 4607872384869556278}
   - component: {fileID: 1654000175361673184}
-  - component: {fileID: -1669536924553513486}
+  - component: {fileID: -4807447924041690560}
   m_Layer: 0
-  m_Name: P_BulletCurve_Boss
+  m_Name: P_BulletHermitCurve_Boss
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -109,7 +109,7 @@ SphereCollider:
   serializedVersion: 3
   m_Radius: 0.5
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &-1669536924553513486
+--- !u!114 &-4807447924041690560
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -118,10 +118,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 1331364910304127374}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 516e811dfaa5aba43b3c853f4f63fc5d, type: 3}
+  m_Script: {fileID: 11500000, guid: 6f3ec18a39d8b8a41809f19171666edf, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   _speed: 10
   _lifeTime: 5
   _damage: 10
-  _damageableType: 3
+  _damageableType: 0
+  StartPoint: {fileID: 0}
+  MidPoint: {fileID: 0}
+  EndPoint: {fileID: 0}
