--- Assets/04. Uyeong/02. Scripts/02-3. Enemy/HermiteBullet.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6f3ec18a39d8b8a41809f19171666edf
